+++ client/src/app/components/SearchInputField.tsx
