+++ client/src/app/components/StatusIcon.tsx
